exoplayer: upd sabr classes

+++ exoplayer-amzn-2.10.6/library/sabr/src/main/java/com/google/android/exoplayer2/source/sabr/parser/models/ConsumedRange.java
@@ -3,10 +3,10 @@ package com.google.android.exoplayer2.source.sabr.parser.models;
 public class ConsumedRange {
     public long startSequenceNumber;
     public long endSequenceNumber;
-    public int startTimeMs;
-    public int durationMs;
+    public long startTimeMs;
+    public long durationMs;
 
-    public ConsumedRange(int startTimeMs, int durationMs, long startSequenceNumber, long endSequenceNumber) {
+    public ConsumedRange(long startTimeMs, long durationMs, long startSequenceNumber, long endSequenceNumber) {
         this.startTimeMs = startTimeMs;
         this.durationMs = durationMs;
         this.startSequenceNumber = startSequenceNumber;
